测试打印功能，fixbug

+++ src/main/java/CSU/Mental/Controller/ProblemAction.java
@@ -264,6 +264,15 @@ public class ProblemAction extends ActionSupport{
 		
 		problem = ProblemService.QueryProblem(pid);
 		
+		if(problem == null) {
+			JSONObject jos = new JSONObject();
+			jos.put("result", "Fail");
+			out.println(jos.toString());
+			out.flush();
+			out.close();
+			return;
+		}
+		
 		flag = ProblemService.DeleteProblem(problem);
 		
 		if(flag) {
